Refactor babel config
Replaced `es2015` preset (which is deprecated) with `env`

Renamed `tests` env to `test` as more widely used and considered as
default in some tools (for example jest).

Added commonjs babel plugin to not treat it as loose. Modules should be
handled as strict as possible.

Upgraded to stage-3 since stage-2 features are not used.